// src/app/dashboard/models/dashboard-states-report.ts
export class DashboardStatesReport{
  actionId: number;
  campaignName: string;
  city: string
  country: string
  countryCode: any;
  deviceType: string;
  emailId: string;
  firstName: string;
  ipAddress: any;
  isp: any;
  lastName: string;
  latitude: string;
  location: string;
  longitude: string;
  os: string;
  state: string;
  subject: string;
  time: any;
  url: string;
  zip: any;
  isShowDetails = false;
  }